Fix subscriptions by replacing deprecated sources API with paymentIntent
For context: for monthly payments we used the now deprecated sources API
in combination with subscriptions. One-time payments already were using
the paymentIntents API.

Now, paymentIntents replace sources for monthly payments as well. This
means we can simplify a lot of the code, since we now use paymentIntents
for everything.

In case of monthly payments, we now create a subscription first without
payment via the new POST /subscription endpoint. Stripe will create the
first invoice for us automatically, and we can return the corresponding
paymentIntent client_secret just like we do on the POST /payment-intent
endpoint.

This means we can treat both endpoints the same on the client-side and
stripe will handle the authorization process for e.g. 3DS enabled credit
cards on its own. Afterwards we POST /payment-intent/finish in both
scenarios.